Debug final do código, fazendo alterações em diversos pontos necessários

+++ tarefa3/src/AjudaExterna.java
@@ -12,7 +12,7 @@ public class AjudaExterna implements Evento {
     }
 
     public void executar(Combatente alvo){
-        int dano = alvo.receberDano(100);
+        double dano = alvo.receberDano(100);
         System.out.printf("\n%s, %s, recebeu %d de dano!\n",alvo.getNome(), Utilidades.verificarClasse(alvo), dano);
     }
 }
